requirements.txt
appnope==0.1.3
asgiref==3.5.0
asttokens==2.0.5
attrs==22.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
decorator==5.1.1
Django==4.0.3
djangorestframework==3.13.1
drf-spectacular==0.23.1
executing==0.8.3
importlib-resources==5.9.0
inflection==0.5.1
ipdb==0.13.9
ipython==8.1.1
jedi==0.18.1
jsonschema==4.9.1
matplotlib-inline==0.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pkgutil-resolve-name==1.3.10
prompt-toolkit==3.0.28
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
pyrsistent==0.18.1
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
toml==0.10.2
traitlets==5.1.1
uritemplate==4.1.1
wcwidth==0.2.5
zipp==3.8.1
